Add rudimentary repository fetching
Splits the input argument into an array on the '/' and further puts the
values from that array into `username` and `repository` variables
respectively.

Also imports Google's go-github package and uses that to fetch repo data
based on the given username/repository, and bails if it can't be found.
For now we just spit out formatted JSON with MarshalIndent, mostly
because `fmt.Printf("%v", json)` is ugly and totally unreadable.

Can't reassign variables in Go so instead of assigning `os.Args[1]` to
`repo` we just use it directly so that `repo` is freed up for the actual
repository data returned from the Github API.