Add list of forks
Gets the list of forks from the API and loops through them printing the
stats out for each repository. Updates the `printRepoStats` function to
use the getters because go-github sets `omitempty` when it fetches the
JSON meaning things like SubscribersCount can be completely unset thus
causing the compiler to bail with a invalid memory error.

"Effective Go" says Go getters shouldn't be prefixed with Get so fuck
knows why Google is doing that, it's ugly as heck especially for things
like `GetHasPages()`... but hey, whatever floats their boat I guess.