Shorten output for stats of fork repos
Reworks the output of forks to have all stats contained within a single
line below the full name of the repository, with a shorter date format.

Shifts the "most recent push" line below the issue count for the long
format so that we can avoid having two separate sets of Printf options.