Prompt for API token (closes #1)
Adds a CLI prompt to paste in the user's Personal Access Token, and
keeps prompting as long as nothing but whitespace is entered.

There's probably a better way to do the for loop, but this works. Would
be nice if we could `while token == "" {}`, but that wouldn't be Go.
Source: https://yourbasic.org/golang/do-while-loop/