Save config to file automatically (closes #3)
Reworks the token prompt so that it saves directly into the config
stuct which saves some hassle as we can then just JSONify `config`
and dump it to a file.

Also adds a way to track when config is loaded so that we don't end up
trying to rewrite the config file every time forkinfo is run.