Allow for overriding auth prompt with flag
Adds a new --no-token flag that bypasses prompting for the Personal
Access Token and automatically runs using the API as a guest.

Since we can now have flags during runtime (whereas previously the
--version flag would end before main()) the native `os.Args` array no
longer cuts the mustard as it also includes the parsed flags, so instead
we switch to using the wrapper functions provided by the flags library.