Add support for use-once tokens
Adds a new --token flag (-t for short) to set a token without
overwriting any existing config such that if the rate limit is ever
reached for one token, a replacement can quickly be used instead.

Config is now loaded slightly sooner to avoid the temporary token being
overwritten when we load the config file, otherwise we'd need to mess
about storing both tokens and extra if statements to check which to use.

Also replaces the configLoaded check with a `saveConfig` check, which is
only set to true if we have to prompt for a token in the first place.