Abstract config functions to their own file
This almost seems too good to be true, as if behind the scenes `go
build` is just concatting or `include`ing the config.go file into the
main forkinfo.go file before it compiles as a single file.

Would normally expect to have to call the config functions some special
way but it seems to all still work as-is... Time will tell.