Show commit stats of forks in output
Adds a line to fork output with status (ahead, behind or diverged) and
number of commits ahead/behind/total. If it errors out (such as a fork
that can't be found - api seems to be cached heavily) then it's removed
from the results, and same goes for those with identical status.