Add option to list commits and their messages
By passing `--show-commits`, it will now list each of the new commits
for the forks that have diverged. Includes shortened SHA hash, author's
name, commited date and the commit message.

Format is getting ugly and somewhat unreadable. Splitting forks with a
line of dashes to have some kind of divider but colours would be better.